TASK-39861 Introduce eXo Server URL parameter in Chat Standalone Deployment mode When sending meeting notes via email, or storing chat meeting in a wiki, the serverBase is used to retrieve data and add link from/to eXo Platform Server. This parameter shouldn't be passed from Client side (Browser) and should be parametered in Chat server side instead.
